requirements.txt
openai>=1.0.0
pypdf>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
